Explicitely convert an arg to the vector<unsigned int> constructor to avoid that it tries to take the constructor that uses two input iterators.
git-svn-id: https://svn.dealii.org/trunk@4791 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/deal.II/source/fe/fe_dgq.cc
@@ -252,7 +252,7 @@ template <int dim>
 std::vector<unsigned int>
 FE_DGQ<dim>::get_dpo_vector(unsigned int deg)
 {
-  std::vector<unsigned int> dpo(dim+1, 0);
+  std::vector<unsigned int> dpo(dim+1, static_cast<unsigned int>(0));
   dpo[dim] = ++deg;
   for (unsigned int i=1;i<dim;++i)
     dpo[dim] *= deg;
